feat: 멤버 별 진단 결과 엔티티 수정 및 세부 진단 결과 관련 DTO 추가

+++ src/main/java/com/taba/nimonaemo/diagnosis/model/entity/DiagnosisResult.java
@@ -36,11 +36,11 @@ public class DiagnosisResult {
 
     @NotNull
     @Column(name = "fine_dead_skin_cells", columnDefinition = "NUMBER(2)")
-    private int findDeadSkinCells;
+    private double findDeadSkinCells;
 
     @NotNull
     @Column(name = "excess_sebum", columnDefinition = "NUMBER(2)")
-    private int excessSebum;
+    private double excessSebum;
 
     @NotNull
     @Column(name = "erythema_between_hair_follicles", columnDefinition = "NUMBER(2)")
